Correct condition when single 'identifier' param
The current condition compares the length of $m to 1, which will never happen as setting $m = $method_map[...] ensures that it will be at least three.

Thus processing drops into "else" clause, generating a PHP error for those methods (get_transaction_details and get_recurring_profile) that should have been captured by the "if" clause.

The proposed solution is not perfect. The 2nd condition is arguably redundant as there are no methods with a single parameter that is not 'identifier'. Conversely it won't deal with these two methods being called with superfluous parameters. However, it would at least avoid errors when they're called correctly.